DP main_svc Значительные изменения

+++ ewm-main-service/src/main/java/ru/practicum/event/dto/EventFullDto.java
@@ -46,5 +46,5 @@ public class EventFullDto {
 
     boolean requestModeration;
 
-    Long views;
+    Integer views;
 }
