+++ ewm-main-service/src/main/java/ru/practicum/event/dto/EventShortDto.java
@@ -28,5 +28,5 @@ public class EventShortDto {
 
     boolean paid;
 
-    Long views;
+    Integer views;
 }
